code refactor for address-space module
