+++ console/console-init/ui/src/modules/address-space/components/AddressSpaceToolbar/AddressSpaceToolbar.tsx
@@ -36,8 +36,6 @@ import { ISearchNameOrNameSpaceAddressSpaceListResponse } from "types/ResponseTy
 import { useApolloClient } from "@apollo/react-hooks";
 import {
   TypeAheadMessage,
-  MAX_ITEM_TO_DISPLAY_IN_TYPEAHEAD_DROPDOWN,
-  NUMBER_OF_RECORDS_TO_DISPLAY_IF_SERVER_HAS_MORE_DATA,
   FetchPolicy,
   TYPEAHEAD_REQUIRED_LENGTH
 } from "constants/constants";
@@ -61,7 +59,7 @@ interface IAddressSpaceListKebabProps {
   onDeleteAll: () => void;
 }
 
-export const AddressSpaceListFilter: React.FunctionComponent<IAddressSpaceListFilterProps> = ({
+export const AddressSpaceToolbar: React.FunctionComponent<IAddressSpaceListFilterProps> = ({
   filterValue,
   setFilterValue,
   filterNames,
