+++ console/console-init/ui/src/modules/address-space/components/EmptyAddressSpace/index.ts
@@ -3,6 +3,4 @@
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
 
-export * from "./CreateAddressSpace";
-export * from "./CreateAddressSpaceConfiguration";
-export * from "./ReviewAddressSpace";
+export * from "./EmptyAddressSpace";
